fix(scheduled-tasks): Separate Redis and SQS
