+++ packages/scheduled-tasks/package.json
@@ -13,22 +13,30 @@
 			"import": "./dist/index.mjs",
 			"require": "./dist/index.js"
 		},
-		"./register": {
-			"import": "./register.mjs",
-			"require": "./register.js"
+		"./register-redis": {
+			"import": "./register-redis.mjs",
+			"require": "./register-redis.js"
+		},
+		"./register-sqs": {
+			"import": "./register-sqs.mjs",
+			"require": "./register-sqs.js"
 		}
 	},
 	"sideEffects": [
-		"./dist/register.js",
-		"./register.js",
-		"./register.mjs"
+		"./dist/register-redis.js",
+		"./dist/register-sqs.js",
+		"./register-redis.js",
+		"./register-redis.mjs",
+		"./register-sqs.js",
+		"./register-sqs.mjs"
 	],
 	"homepage": "https://github.com/sapphiredev/plugins/tree/main/packages/scheduled-tasks",
 	"scripts": {
 		"test": "jest",
 		"lint": "eslint src tests --ext ts --fix",
-		"build": "tsc -b src && yarn esm:register && yarn esm:default",
-		"esm:register": "gen-esm-wrapper dist/register.js dist/register.mjs",
+		"build": "tsc -b src && yarn esm:register-redis && yarn esm:register-sqs && yarn esm:default",
+		"esm:register-redis": "gen-esm-wrapper dist/register-redis.js dist/register-redis.mjs",
+		"esm:register-sqs": "gen-esm-wrapper dist/register-sqs.js dist/register-sqs.mjs",
 		"esm:default": "gen-esm-wrapper dist/index.js dist/index.mjs",
 		"prepublishOnly": "yarn build"
 	},
@@ -36,7 +44,7 @@
 		"@sapphire/time-utilities": "^1.3.8",
 		"tslib": "^2.3.1"
 	},
-	"peerDependencies": {
+	"devDependencies": {
 		"@types/bull": "^3.15.5",
 		"bull": "^3.29.2",
 		"sqs-consumer": "^5.6.0",
@@ -50,7 +58,8 @@
 	"files": [
 		"dist",
 		"!dist/*.tsbuildinfo",
-		"register.*"
+		"register-redis.*",
+		"register-sqs.*"
 	],
 	"engines": {
 		"node": ">=v14.18.0",
