+++ packages/scheduled-tasks/register-redis.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/register-redis';
