+++ packages/scheduled-tasks/register-redis.js
@@ -1,0 +1,4 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const tslib = require('tslib');
+
+tslib.__exportStar(require('./dist/register-redis.js'), exports);
