+++ packages/scheduled-tasks/register-redis.mjs
@@ -1,0 +1,1 @@
+export * from './dist/register-redis.mjs';
