+++ packages/scheduled-tasks/register-sqs.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/register-sqs';
