+++ packages/scheduled-tasks/register-sqs.mjs
@@ -1,0 +1,1 @@
+export * from './dist/register-sqs.mjs';
