--- packages/scheduled-tasks/register.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/register';
