--- packages/scheduled-tasks/register.mjs
@@ -1,1 +1,0 @@
-export * from './dist/register.mjs';
