--- packages/scheduled-tasks/src/lib/strategies/index.ts
@@ -1,2 +1,0 @@
-export * from './ScheduledTaskRedisStrategy';
-export * from './ScheduledTaskSQSStrategy';
