--- packages/scheduled-tasks/src/lib/types/ScheduledTaskStrategy.ts
@@ -1,3 +1,0 @@
-import type { ScheduledTaskRedisStrategy, ScheduledTaskSQSStrategy } from '../strategies';
-
-export type ScheduledTaskStrategy = ScheduledTaskRedisStrategy | ScheduledTaskSQSStrategy;
