+++ packages/scheduled-tasks/src/lib/types/index.ts
@@ -2,4 +2,3 @@ export * from './ScheduledTaskBaseStrategy';
 export * from './ScheduledTaskCreateRepeatedTask';
 export * from './ScheduledTasksOptions';
 export * from './ScheduledTasksTaskOptions';
-export * from './ScheduledTaskStrategy';
