+++ packages/scheduled-tasks/src/register-redis.ts
@@ -1,0 +1,3 @@
+import './index';
+
+export * from './lib/strategies/ScheduledTaskRedisStrategy';
