+++ packages/scheduled-tasks/src/register-sqs.ts
@@ -1,0 +1,3 @@
+import './index';
+
+export * from './lib/strategies/ScheduledTaskSQSStrategy';
